+++ src/main/java/com/malikh/StockWatchdog/Service/StockService.java
@@ -3,6 +3,8 @@ package com.malikh.StockWatchdog.Service;
 import java.util.List;
 import java.util.Optional;
 
+import javax.management.RuntimeErrorException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,8 +13,12 @@ import com.malikh.StockWatchdog.Repository.StockRepository;
 
 @Service
 public class StockService {
-    @Autowired
-    public StockRepository stockRepo;
+    private final StockRepository stockRepo;
+
+    public StockService(StockRepository stockRepo) {
+        this.stockRepo = stockRepo;
+    }
+
     // CRUD
     // Create
     public void createStock(Stock s){
@@ -27,12 +33,18 @@ public class StockService {
         return stockRepo.findById(s);
     }
     //UPDATE
-    public void updateStock(String s, Stock st){
-        st.setSymbol(s);
-        stockRepo.save(st);
+    public Stock updateStock(String s, Stock updatedStock){
+        Stock existingStock = stockRepo.findById(s)
+                .orElseThrow(() -> new RuntimeErrorException("Stock not found"));
+        existingStock.setCompanyName(updatedStock.getCompanyName());
+        existingStock.setPrice(updatedStock.getPrice());
+        return stockRepo.save(existingStock);
     }
     //DELETE
     public void deleteStock(String s){
+        if (!stockRepo.existsById(s)) {
+            throw new RuntimeErrorException("Stock not found");
+        }
         stockRepo.deleteById(s);
     }
